+++ server/src/prizmServices.ts
@@ -135,7 +135,16 @@ const serviceDescriptions = {
     },
     pccis: {
         name: 'PCCIS',
-        logFile: /Pccis\d+\/.*/i
+        logFile: /Pccis\d+\/ImagingServices.log/i,
+        requestPath: /\/PCCIS\/V1\/(Page|Document|ViewingSession).*/i
+    },
+    loadBalancer: {
+        name: 'LoadBalancer',
+        logFile: 'plb.sep_single.log'
+    },
+    pas: {
+        name: 'PAS',
+        logFile: /pas\/pas-\d+.log/i
     }
 };