+++ server/test/codeNavigator.spec.ts
@@ -213,6 +213,173 @@ describe('CodeNavigator', () => {
                     });
                 });
             });
+
+            describe('should return location when log entry describes request pccis -> rcs', () => {
+
+                it('for PCCIS log entry', () => {
+                    const logItem = { 
+                        "time": "2018-03-16T14:08:12.948Z", 
+                        "gid": "5LWSCTeH9bzPB7HHGJI03A", 
+                        "name": "PCCIS", "level": 30, 
+                        "taskid": 59, "pid": 2418, "tid": 121, 
+                        "msg": "InternalRequest (RCS)", 
+                        "reqBegin": true, 
+                        "req": { 
+                            "method": "POST", 
+                            "path": "/documentAttributes", 
+                            "port": 19005 
+                        } 
+                    };
+                    const definition = codeNavigator.getDefinition(logItem);
+                    expect(definition).eql({
+                        uri: 'RasterConversionService.log',
+                        range: {
+                            start: { line: 0, character: 0 },
+                            end: { line: 0, character: 339 }
+                        }
+                    });
+                });
+
+                it('for RasterConversionService log entry', () => {
+                    const logItem = { 
+                        "gid": "5LWSCTeH9bzPB7HHGJI03A", 
+                        "name": "RCS", "time": "2018-03-16T14:08:12.950Z", 
+                        "pid": 2143, "level": 30, "tid": 2147, "taskid": 20, 
+                        "taskBegin": true, 
+                        "taskName": "Request", 
+                        "parent": { 
+                            "name": "PCCIS", 
+                            "pid": 2418, 
+                            "taskid": 59 
+                        }, 
+                        "reqAccepted": true, 
+                        "req": { 
+                            "method": "POST", 
+                            "port": 19005, 
+                            "path": "/RCS/documentAttributes" 
+                        } 
+                    };
+                    const definition = codeNavigator.getDefinition(logItem);
+                    expect(definition).eql({
+                        uri: 'Pccis2/ImagingServices.log',
+                        range: {
+                            start: { line: 2, character: 0 },
+                            end: { line: 2, character: 265 }
+                        }
+                    });
+                });
+            });
+
+            describe('should return location when log entry describes request plb -> pccis', () => {
+
+                it('for LoadBalancer log entry', () => {
+                    const logItem = { 
+                        "name": "LoadBalancer", 
+                        "hostname": "PTPcanaryUbuntu14CoreWin10ChromeJSP20180375095316316s13205451", 
+                        "pid": 2562, "taskid": 100, "gid": "MPvux5kbwGfis8K47M9gVA", "level": 30, 
+                        "reqBegin": true, 
+                        "req": { 
+                            "method": "GET", 
+                            "url": "http://localhost:19002/PCCIS/V1/Page/q/0/Tile/0/1020/1028/260?DocumentID=ueiLcFo6dudAv59Bifi04Ku7-4_nrtVIGCED5UM8HiM2jSiol0ec9J3qx_1XNh3jU3wVbAIifKhSN06CdOarrRo8VzpiRlsdTV-bJccq-_gikwnvq0GoTIPfD-6PyOtj_i8PUHtRee3NI-iwBFZCrhA&Scale=1&ContentType=png&Quality=100" 
+                        }, 
+                        "msg": "", 
+                        "time": "2018-03-16T14:08:13.470Z", "v": 0 
+                    };
+                    const definition = codeNavigator.getDefinition(logItem);
+                    expect(definition).eql({
+                        uri: 'Pccis2/ImagingServices.log',
+                        range: {
+                            start: { line: 6, character: 0 },
+                            end: { line: 6, character: 364 }
+                        }
+                    });
+                });
+
+                it('for PCCIS log entry', () => {
+                    const logItem = { 
+                        "time": "2018-03-16T14:08:13.471Z", "gid": "MPvux5kbwGfis8K47M9gVA", 
+                        "name": "PCCIS", "level": 30, "taskid": 66, "pid": 2418, "tid": 87, 
+                        "msg": "ProcessRequest", 
+                        "taskBegin": true, 
+                        "parent": { 
+                            "name": "LoadBalancer", 
+                            "pid": 2562, 
+                            "taskid": 100 
+                        }, 
+                        "reqAccepted": true, 
+                        "req": { 
+                            "method": "GET", 
+                            "path": "/PCCIS/V1/Page/q/0/Tile/0/1020/1028/260", 
+                            "port": 19002 
+                        } 
+                    };
+                    const definition = codeNavigator.getDefinition(logItem);
+                    expect(definition).eql({
+                        uri: 'plb.sep_single.log',
+                        range: {
+                            start: { line: 3, character: 0 },
+                            end: { line: 3, character: 549 }
+                        }
+                    });
+                });
+            });
+
+            describe('should return location when log entry describes request pas -> plb -> pccis', () => {
+
+                it('for LoadBalancer log entry', () => {
+                    const logItem = { 
+                        "name": "LoadBalancer", 
+                        "hostname": "PTPcanaryUbuntu14CoreWin10ChromeJSP20180375095316316s13205451", 
+                        "pid": 2562, "taskid": 100, 
+                        "gid": "MPvux5kbwGfis8K47M9gVA", "level": 30, 
+                        "taskBegin": true, 
+                        "parent": { 
+                            "name": "PAS", 
+                            "pid": 3353, 
+                            "taskid": 22 
+                        }, 
+                        "reqAccepted": true, 
+                        "req": { 
+                            "method": "GET", 
+                            "path": "/PCCIS/V1/Page/q/0/Tile/0/1020/1028/260?DocumentID=ueiLcFo6dudAv59Bifi04Ku7-4_nrtVIGCED5UM8HiM2jSiol0ec9J3qx_1XNh3jU3wVbAIifKhSN06CdOarrRo8VzpiRlsdTV-bJccq-_gikwnvq0GoTIPfD-6PyOtj_i8PUHtRee3NI-iwBFZCrhA&Scale=1&Quality=100&ContentType=png", 
+                            "port": 3000 
+                        }, 
+                        "msg": "", "time": "2018-03-16T14:08:13.470Z", "v": 0 
+                    };
+                    const definition = codeNavigator.getDefinition(logItem);
+                    expect(definition).eql({
+                        uri: 'pas/pas-1.log',
+                        range: {
+                            start: { line: 1, character: 0 },
+                            end: { line: 1, character: 639 }
+                        }
+                    });
+                });
+
+                it('for PAS log entry', () => {
+                    const logItem = { 
+                        "name": "PAS", 
+                        "hostname": "PTPcanaryUbuntu14CoreWin10ChromeJSP20180375095316316s13205451", 
+                        "pid": 3353, "taskid": 22, "gid": "MPvux5kbwGfis8K47M9gVA", "level": 30, 
+                        "reqBegin": true, "operation": "proxyToPccis", 
+                        "req": { 
+                            "method": "GET", 
+                            "baseUrl": "http://localhost:18681", 
+                            "path": "/PCCIS/V1/Page/q/0/Tile/0/1020/1028/260", 
+                            "qs": { "DocumentID": "ueiLcFo6dudAv59Bifi04Ku7-4_nrtVIGCED5UM8HiM2jSiol0ec9J3qx_1XNh3jU3wVbAIifKhSN06CdOarrRo8VzpiRlsdTV-bJccq-_gikwnvq0GoTIPfD-6PyOtj_i8PUHtRee3NI-iwBFZCrhA", "Scale": "1", "Quality": "100", "ContentType": "png" } 
+                        }, 
+                        "timeline": "start", "msg": "", "time": "2018-03-16T14:08:13.469Z", "v": 0 
+                    };
+                    const definition = codeNavigator.getDefinition(logItem);
+                    expect(definition).eql({
+                        uri: 'Pccis2/ImagingServices.log',
+                        range: {
+                            start: { line: 6, character: 0 },
+                            end: { line: 6, character: 364 }
+                        }
+                    });
+                });
+            });
         });
     });