+++ client/CHANGELOG.md
@@ -2,6 +2,7 @@
 - Improve PCCIS support
 - Improve PAS support
 - Improve CodeLens performance
+- Added navigation to the long operation end
 
 # Version 0.0.2
 - CodeLens for all tasks