+++ client/CHANGELOG.md
@@ -1,3 +1,5 @@
+# Version 0.0.4
+
 # Version 0.0.3
 - Improve PCCIS support
 - Improve PAS support